pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "compliance-agent"
version = "2.0.0"
description = "Enhanced compliance agent - multi-document orchestration with parallel execution"
requires-python = ">=3.12"
dependencies = [
    "langgraph",
    "langchain",
    "langchain-community", 
    "langchain-openai",
    "langchain-core",
    "openai",
    "anthropic",
    "httpx",
    "python-dotenv",
    "pydantic",
    "python-dateutil",
    "aiohttp"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "mypy", 
    "ruff"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
